as u16 in value()

+++ fancy-garbling/src/wire.rs
@@ -91,7 +91,7 @@ impl Modulus {
     pub fn value(&self) -> u16 {
         match self {
             Modulus::Zq { q } => *q,
-            Modulus::GF4 { p } => *p ,
+            Modulus::GF4 { p } => *p as u16 ,
         }
     }
